ISS83 - refactor mutator class Speed
