+++ python_src/models/mutator/__init__.py
@@ -1,1 +1,3 @@
 from .mutator import Mutator, categorize_mutator, CAT_A, CAT_B
+from .mutatorA import MutatorTypeA
+
